isolate type folders
