+++ client/api/command-types.go
@@ -1,0 +1,11 @@
+package api
+
+type CallIssueCommandResponse struct {
+	StatusCode int
+	Body       string
+}
+
+type ResponseBody struct {
+	Message string
+	Success bool
+}
