+++ client/vehicle-state/vehicle-state.go
@@ -10,34 +10,6 @@ import (
 	"tesla-app/client/helpers"
 )
 
-type VehicleStateResponse struct {
-	State string
-	Vin   string
-}
-
-type TeslaVehicleApiResponse struct {
-	Response struct {
-		ID             int64  `json:"id"`
-		UserID         int64  `json:"user_id"`
-		VehicleID      int64  `json:"vehicle_id"`
-		Vin            string `json:"vin"`
-		Color          any    `json:"color"`
-		AccessType     string `json:"access_type"`
-		GranularAccess struct {
-			HidePrivate bool `json:"hide_private"`
-		} `json:"granular_access"`
-		Tokens                 any    `json:"tokens"`
-		State                  string `json:"state"`
-		InService              bool   `json:"in_service"`
-		IDS                    string `json:"id_s"`
-		CalendarEnabled        bool   `json:"calendar_enabled"`
-		APIVersion             int    `json:"api_version"`
-		BackseatToken          any    `json:"backseat_token"`
-		BackseatTokenUpdatedAt any    `json:"backseat_token_updated_at"`
-		BleAutopairEnrolled    bool   `json:"ble_autopair_enrolled"`
-	} `json:"response"`
-}
-
 func VehicleState() (*VehicleStateResponse, error) {
 	tokenStore, state := helpers.GetTokenStore()
 	carId := os.Getenv("MY_CAR_ID")
