+++ client/vehicle-state/wake-types.go
@@ -1,0 +1,11 @@
+package vehicle
+
+type WakeResponse struct {
+	State string
+}
+
+type TeslaVehicleWakeResponse struct {
+	Response struct {
+		State string `json:"state"`
+	}
+}
