--- public/tfetch-auth/.prettierrc
@@ -1,11 +1,0 @@
-{
-  "semi": true,
-  "singleQuote": true,
-  "tabWidth": 2,
-  "useTabs": false,
-  "trailingComma": "es5",
-  "bracketSpacing": true,
-  "arrowParens": "always",
-  "printWidth": 80,
-  "endOfLine": "lf"
-}
