--- public/tfetch-auth/index.html
@@ -1,47 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link href="./output.css" rel="stylesheet" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Tesla Fetch</title>
-  </head>
-
-  <body class="bg-stone-950 p-20">
-    <div id="app" class="flex items-center w-full">
-      <div class="mr-3">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 100 100"
-          width="42"
-          height="42"
-        >
-          <circle
-            cx="50"
-            cy="50"
-            r="45"
-            stroke="#4CAF50"
-            stroke-width="5"
-            fill="none"
-          />
-          <path
-            d="M30 50 L45 65 L70 35"
-            stroke="#4CAF50"
-            stroke-width="5"
-            fill="none"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-          />
-        </svg>
-      </div>
-      <div>
-        <div>
-          <p class="text-stone-50 text-lg">Auth Successful</p>
-        </div>
-        <div>
-          <p class="text-stone-50">You now have access to your vehicle data</p>
-        </div>
-      </div>
-    </div>
-  </body>
-</html>
