--- public/tfetch-auth/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "tfetch-auth",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "tailwind": "npx tailwindcss -i ./input.css -o ./output.css --watch"
-  },
-  "devDependencies": {
-    "tailwind": "^4.0.0",
-    "vite": "^6.0.5"
-  },
-  "dependencies": {
-    "prettier": "^3.4.2"
-  }
-}
