--- public/tfetch-auth/tailwind.config.js
@@ -1,4 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: ['./*.html'],
-};
